Try to restore C bindings as a separate library build.
