+++ CMakeLists.txt
@@ -49,6 +49,7 @@ option(MANIFOLD_PAR "Enable Parallel backend" OFF)
 option(MANIFOLD_TEST "Enable testing suite" ON)
 option(BUILD_SHARED_LIBS "Build shared library" ON)
 include(CMakeDependentOption)
+cmake_dependent_option(MANIFOLD_CBIND "Build C (FFI) bindings" ON "MANIFOLD_CROSS_SECTION" OFF)
 option(MANIFOLD_PYBIND "Build python bindings" OFF)
 cmake_dependent_option(MANIFOLD_JSBIND "Build js binding" ON "EMSCRIPTEN" OFF)
 
@@ -301,6 +302,7 @@ message(STATUS "MANIFOLD_EXPORT:             ${MANIFOLD_EXPORT}")
 message(STATUS "MANIFOLD_TEST:               ${MANIFOLD_TEST}")
 message(STATUS "MANIFOLD_FUZZ:               ${MANIFOLD_FUZZ}")
 message(STATUS "MANIFOLD_DEBUG:              ${MANIFOLD_DEBUG}")
+message(STATUS "MANIFOLD_CBIND:              ${MANIFOLD_CBIND}")
 message(STATUS "MANIFOLD_PYBIND:             ${MANIFOLD_PYBIND}")
 message(STATUS "MANIFOLD_JSBIND:             ${MANIFOLD_JSBIND}")
 message(STATUS "MANIFOLD_EXCEPTIONS:         ${MANIFOLD_EXCEPTIONS}")
