+++ bindings/c/cross.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../conv.h"
+#include "./conv.h"
 #include "manifold/common.h"
 #include "manifold/cross_section.h"
 #include "manifold/types.h"
