+++ bindings/c/rect.cpp
